basic-publish-and-subscribe-py/pyproject.toml
[tool.poetry]
name = "basic-publish-and-subscribe-py"
version = "0.1.0"
description = ""
authors = ["kzmake <kzmake.i3a@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
googleapis-common-protos = "^1.53.0"
grpcio = "^1.39.0"
grpclib = "^0.4.2"
ulid-py = "^1.1.0"
dapr = "^1.2.0"
dapr-ext-grpc = "^1.2.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"
grpcio-tools = "^1.39.0"
grpclib = "^0.4.2"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
